HHH-18139 clean up / reorg of generator binding related code
