+++ hibernate-core/src/main/java/org/hibernate/boot/model/IdentifierGeneratorDefinition.java
@@ -6,26 +6,27 @@
  */
 package org.hibernate.boot.model;
 
-import java.io.Serializable;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
-
+import jakarta.persistence.GenerationType;
 import org.hibernate.AnnotationException;
 import org.hibernate.AssertionFailure;
 import org.hibernate.Internal;
-import org.hibernate.boot.internal.GenerationStrategyInterpreter;
 import org.hibernate.boot.models.annotations.internal.SequenceGeneratorJpaAnnotation;
 import org.hibernate.boot.models.annotations.internal.TableGeneratorJpaAnnotation;
 import org.hibernate.id.IdentifierGenerator;
 import org.hibernate.internal.util.StringHelper;
+import org.hibernate.models.spi.TypeDetails;
 
-import jakarta.persistence.GenerationType;
+import java.io.Serializable;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
 
 import static java.util.Collections.emptyMap;
+import static java.util.Collections.singletonMap;
 import static java.util.Collections.unmodifiableMap;
-import static org.hibernate.boot.internal.GenerationStrategyInterpreter.STRATEGY_INTERPRETER;
+import static org.hibernate.boot.model.internal.GeneratorParameters.interpretSequenceGenerator;
+import static org.hibernate.boot.model.internal.GeneratorParameters.interpretTableGenerator;
+import static org.hibernate.boot.model.internal.GeneratorStrategies.generatorStrategy;
 import static org.hibernate.boot.models.JpaAnnotations.SEQUENCE_GENERATOR;
 import static org.hibernate.boot.models.JpaAnnotations.TABLE_GENERATOR;
 import static org.hibernate.internal.util.collections.CollectionHelper.isEmpty;
@@ -93,7 +94,7 @@ public class IdentifierGeneratorDefinition implements Serializable {
 	@Internal
 	public static IdentifierGeneratorDefinition createImplicit(
 			String name,
-			Class<?> idType,
+			TypeDetails idType,
 			String generatorName,
 			GenerationType generationType) {
 		// If we were unable to locate an actual matching named generator assume
@@ -104,49 +105,29 @@ public class IdentifierGeneratorDefinition implements Serializable {
 		}
 
 		switch ( generationType ) {
-			case SEQUENCE: {
+			case SEQUENCE:
 				return buildSequenceGeneratorDefinition( name );
-			}
-			case TABLE: {
+			case TABLE:
 				return buildTableGeneratorDefinition( name );
-			}
-			case IDENTITY: {
+			case IDENTITY:
 				throw new AnnotationException(
 						"@GeneratedValue annotation specified 'strategy=IDENTITY' and 'generator'"
 								+ " but the generator name is unnecessary"
 				);
-			}
-			case UUID: {
+			case UUID:
 				throw new AnnotationException(
 						"@GeneratedValue annotation specified 'strategy=UUID' and 'generator'"
 								+ " but the generator name is unnecessary"
 				);
-			}
-			case AUTO: {
-				final String strategyName = STRATEGY_INTERPRETER.determineGeneratorName(
-						generationType,
-						new GenerationStrategyInterpreter.GeneratorNameDeterminationContext() {
-							@Override
-							public Class<?> getIdType() {
-								return idType;
-							}
-							@Override
-							public String getGeneratedValueGeneratorName() {
-								return generatorName;
-							}
-						}
-				);
-
+			case AUTO:
 				return new IdentifierGeneratorDefinition(
 						name,
-						strategyName,
-						Collections.singletonMap( IdentifierGenerator.GENERATOR_NAME, name )
+						generatorStrategy( generationType, generatorName, idType ),
+						singletonMap( IdentifierGenerator.GENERATOR_NAME, name )
 				);
-			}
+			default:
+				throw new AssertionFailure( "unknown generator type: " + generationType );
 		}
-
-		throw new AssertionFailure( "unknown generator type: " + generationType );
-
 	}
 
 	private static IdentifierGeneratorDefinition buildTableGeneratorDefinition(String name) {
@@ -155,7 +136,7 @@ public class IdentifierGeneratorDefinition implements Serializable {
 		if ( StringHelper.isNotEmpty( name ) ) {
 			tableGeneratorUsage.name( name );
 		}
-		STRATEGY_INTERPRETER.interpretTableGenerator( tableGeneratorUsage, builder );
+		interpretTableGenerator( tableGeneratorUsage, builder );
 		return builder.build();
 	}
 
@@ -165,7 +146,7 @@ public class IdentifierGeneratorDefinition implements Serializable {
 		if ( StringHelper.isNotEmpty( name ) ) {
 			sequenceGeneratorUsage.name( name );
 		}
-		STRATEGY_INTERPRETER.interpretSequenceGenerator( sequenceGeneratorUsage, builder );
+		interpretSequenceGenerator( sequenceGeneratorUsage, builder );
 		return builder.build();
 	}
