+++ hibernate-core/src/main/java/org/hibernate/boot/model/internal/CollectionBinder.java
@@ -1613,7 +1613,7 @@ public abstract class CollectionBinder {
 	}
 
 	TypeDetails getElementType() {
-		if ( isDefault( targetEntity, buildingContext ) ) {
+		if ( isDefault( targetEntity ) ) {
 			if ( collectionElementType != null ) {
 				return collectionElementType;
 			}
