+++ hibernate-core/src/main/java/org/hibernate/boot/model/internal/IdGeneratorResolverSecondPass.java
@@ -9,7 +9,6 @@ package org.hibernate.boot.model.internal;
 import java.util.Map;
 
 import org.hibernate.MappingException;
-import org.hibernate.boot.model.IdentifierGeneratorDefinition;
 import org.hibernate.boot.spi.MetadataBuildingContext;
 import org.hibernate.boot.spi.SecondPass;
 import org.hibernate.mapping.PersistentClass;
@@ -27,7 +26,7 @@ public class IdGeneratorResolverSecondPass implements SecondPass {
 	private final String generatorType;
 	private final String generatorName;
 	private final MetadataBuildingContext buildingContext;
-	private IdentifierGeneratorDefinition localIdentifierGeneratorDefinition;
+//	private IdentifierGeneratorDefinition localIdentifierGeneratorDefinition;
 
 	public IdGeneratorResolverSecondPass(
 			SimpleValue id,
@@ -42,19 +41,33 @@ public class IdGeneratorResolverSecondPass implements SecondPass {
 		this.buildingContext = buildingContext;
 	}
 
-	public IdGeneratorResolverSecondPass(
-			SimpleValue id,
-			MemberDetails idAttributeMember,
-			String generatorType,
-			String generatorName,
-			MetadataBuildingContext buildingContext,
-			IdentifierGeneratorDefinition localIdentifierGeneratorDefinition) {
-		this( id, idAttributeMember, generatorType, generatorName, buildingContext );
-		this.localIdentifierGeneratorDefinition = localIdentifierGeneratorDefinition;
-	}
+//	public IdGeneratorResolverSecondPass(
+//			SimpleValue id,
+//			MemberDetails idAttributeMember,
+//			String generatorType,
+//			String generatorName,
+//			MetadataBuildingContext buildingContext,
+//			IdentifierGeneratorDefinition localIdentifierGeneratorDefinition) {
+//		this( id, idAttributeMember, generatorType, generatorName, buildingContext );
+//		this.localIdentifierGeneratorDefinition = localIdentifierGeneratorDefinition;
+//	}
 
 	@Override
 	public void doSecondPass(Map<String, PersistentClass> idGeneratorDefinitionMap) throws MappingException {
-		makeIdGenerator( id, idAttributeMember, generatorType, generatorName, buildingContext, localIdentifierGeneratorDefinition );
+//		makeIdGenerator( id, idAttributeMember, generatorType, generatorName, buildingContext, localIdentifierGeneratorDefinition );
+		makeIdGenerator( id, idAttributeMember, generatorType, generatorName, buildingContext, null );
 	}
+
+//	public static void makeIdGenerator(
+//			SimpleValue id,
+//			MemberDetails idAttributeMember,
+//			String generatorType,
+//			String generatorName,
+//			MetadataBuildingContext buildingContext,
+//			IdentifierGeneratorDefinition foreignKGeneratorDefinition) {
+//		makeIdGenerator( id, idAttributeMember, generatorType, generatorName, buildingContext,
+//				foreignKGeneratorDefinition != null
+//						? Map.of( foreignKGeneratorDefinition.getName(), foreignKGeneratorDefinition )
+//						: null );
+//	}
 }
