+++ hibernate-core/src/main/java/org/hibernate/internal/util/collections/CollectionHelper.java
@@ -6,7 +6,6 @@
  */
 package org.hibernate.internal.util.collections;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
